pyproject.toml
[project]
name = "ire"
version = "0.1.0"
description = "IRE (Ingredient-to-Recipe Engine) - Food ingredient detection system"
authors = [
    {name = "tuong.nguyen",email = "richardnguyen0715@gmail.com"}
]
readme = "README.md"
requires-python = "<3.15,>=3.10"
dependencies = [
    "jupyter (>=1.1.1,<2.0.0)",
    "roboflow (>=1.2.16,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "ultralytics (>=8.4.21,<9.0.0)",
    "torch (>=2.10.0,<3.0.0)",
    "gradio (>=6.9.0,<7.0.0)",
    "psutil (>=7.2.2,<8.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
